ahx_encoder: fixed subband allocation values
So it turns out that ahx files use specific values for their bit
allocation values. It seems to be hard coded in every ahx file (At least
for those in SA2). I removed the if else statements and replaced them
with a table for cleaner and more optimal code. Also had to edit the
bit allocation writing to correspond with the new values.

I'll have to add proper scfsi code sometime.